refactor(ui-dashboard): split oracle chart + restrict to median samples
Two more findings from the PR-624 review:

- Split the oversized chart module. AGENTS.md soft-caps files at 600
  lines; oracle-chart.tsx grew to 683 in this PR. Extracted the wheel
  handler + its zoom helpers to `oracle-chart-wheel.ts` and the hover
  formatter / number-format helper to `oracle-chart-hover.ts`. Main
  file is now 551 lines.

- Restrict the chart to `oracle_median_updated` samples. BreakerBox
  only evaluates trip conditions on MedianUpdated — a single outlier
  reporter that doesn't move the median never trips a breaker. The
  prior query mixed `oracle_reported` rows in, so a per-reporter
  outlier could render as a red "would trip" point even though the
  median (and the breaker) stayed inside the band.

- Bump the size-limit budget from 1830 KB → 1850 KB. The breaker
  plumbing adds ~3 KB brotli; 1850 KB gives ~1.4 % headroom over
  current so this and small follow-ons fit without re-bumping.
